# Rewrite 4x4 det as product of 2x2 minors for improved performance